Convert the PWM duty cylcle to double. Preserve old setters, as Double can't be directly assigned from int constants (oppsed to double)

+++ pi4j-core/src/main/java/com/pi4j/io/pwm/Pwm.java
@@ -237,7 +237,7 @@ public interface Pwm extends IO<Pwm, PwmConfig, PwmProvider>, OnOff<Pwm> {
      * @return duty-cycle value expressed as a percentage (rage: 0-100)
      * @throws IOException if fails to communicate with the PWM pin
      */
-    default Integer dutyCycle() throws IOException {
+    default double dutyCycle() throws IOException {
         return getDutyCycle();
     }
